move helper methods to also use it for the rdf parser
