+++ src/spdx/parser/jsonlikedict/annotation_parser.py
@@ -15,8 +15,8 @@ from spdx.model.actor import Actor
 from spdx.model.annotation import Annotation, AnnotationType
 from spdx.parser.error import SPDXParsingError
 from spdx.parser.jsonlikedict.actor_parser import ActorParser
-from spdx.parser.jsonlikedict.dict_parsing_functions import construct_or_raise_parsing_error, \
-    parse_field_or_log_error, append_parsed_field_or_log_error, raise_parsing_error_if_logger_has_messages
+from spdx.parser.jsonlikedict.dict_parsing_functions import parse_field_or_log_error, append_parsed_field_or_log_error
+from spdx.parser.parsing_functions import construct_or_raise_parsing_error, raise_parsing_error_if_logger_has_messages
 from spdx.datetime_conversions import datetime_from_str
 from spdx.parser.logger import Logger
