+++ src/spdx/parser/jsonlikedict/dict_parsing_functions.py
@@ -12,23 +12,15 @@ from typing import Any, Callable, Dict, List, Optional
 
 from spdx.model.spdx_no_assertion import SpdxNoAssertion
 from spdx.model.spdx_none import SpdxNone
-from common.typing.constructor_type_errors import ConstructorTypeErrors
 from spdx.parser.error import SPDXParsingError
 from spdx.parser.logger import Logger
+from spdx.parser.parsing_functions import raise_parsing_error_if_logger_has_messages
 
 
 def json_str_to_enum_name(json_str: str) -> str:
     if not isinstance(json_str, str):
         raise SPDXParsingError([f"Type for enum must be str not {type(json_str).__name__}"])
     return json_str.replace("-", "_").upper()
-
-
-def construct_or_raise_parsing_error(object_to_construct: Any, args_for_construction: Dict) -> Any:
-    try:
-        constructed_object = object_to_construct(**args_for_construction)
-    except ConstructorTypeErrors as err:
-        raise SPDXParsingError([f"Error while constructing {object_to_construct.__name__}: {err.get_messages()}"])
-    return constructed_object
 
 
 def parse_field_or_log_error(logger: Logger, field: Any, parsing_method: Callable = lambda x: x, default: Any = None,
@@ -57,14 +49,6 @@ def append_parsed_field_or_log_error(logger: Logger, list_to_append_to: List[Any
     except (TypeError, ValueError) as err:
         logger.extend(err.args[0])
     return list_to_append_to
-
-
-def raise_parsing_error_if_logger_has_messages(logger: Logger, parsed_object_name: str = None):
-    if logger.has_messages():
-        if parsed_object_name:
-            raise SPDXParsingError([f"Error while parsing {parsed_object_name}: {logger.get_messages()}"])
-        else:
-            raise SPDXParsingError(logger.get_messages())
 
 
 def parse_field_or_no_assertion_or_none(field: Optional[str], method_for_field: Callable = lambda x: x) -> Any:
