+++ src/spdx/parser/jsonlikedict/license_expression_parser.py
@@ -15,8 +15,8 @@ from license_expression import LicenseExpression, Licensing, ExpressionError
 from spdx.model.spdx_no_assertion import SpdxNoAssertion
 from spdx.model.spdx_none import SpdxNone
 from spdx.parser.error import SPDXParsingError
-from spdx.parser.jsonlikedict.dict_parsing_functions import append_parsed_field_or_log_error, \
-    raise_parsing_error_if_logger_has_messages
+from spdx.parser.jsonlikedict.dict_parsing_functions import append_parsed_field_or_log_error
+from spdx.parser.parsing_functions import raise_parsing_error_if_logger_has_messages
 from spdx.parser.logger import Logger
