+++ src/spdx/parser/jsonlikedict/snippet_parser.py
@@ -16,8 +16,9 @@ from spdx.model.snippet import Snippet
 from spdx.model.spdx_no_assertion import SpdxNoAssertion
 from spdx.model.spdx_none import SpdxNone
 from spdx.parser.error import SPDXParsingError
-from spdx.parser.jsonlikedict.dict_parsing_functions import construct_or_raise_parsing_error, parse_field_or_log_error, \
+from spdx.parser.jsonlikedict.dict_parsing_functions import parse_field_or_log_error, \
     parse_field_or_no_assertion_or_none
+from spdx.parser.parsing_functions import construct_or_raise_parsing_error
 
 from spdx.parser.jsonlikedict.license_expression_parser import LicenseExpressionParser
 from spdx.parser.logger import Logger
